// src/types/Match.ts
export type MatchResult = {
    homeGoals: number;
    awayGoals: number;
    matchEndResult: string;
};

export type Match = {
    matchId: number;
    matchNo: string;
    activityAreaId: number;
    activityAreaLatitude: string;
    activityAreaLongitude: string;
    activityAreaName: string;
    activityAreaNo: string;
    admOrgId: number;
    arrOrgId: number;
    arrOrgNo: string;
    arrOrgName: string;
    awayteamId: number;
    awayteamOrgNo: string;
    awayteam: string;
    awayteamOrgName: string;
    awayteamOverriddenName: string | null;
    awayteamClubOrgId: number;
    hometeamId: number;
    hometeam: string;
    hometeamOrgName: string;
    hometeamOverriddenName: string | null;
    hometeamOrgNo: string;
    hometeamClubOrgId: number;
    roundId: number;
    roundName: string;
    seasonId: number;
    tournamentName: string;
    tournamentId: number;
    matchDate: string;
    matchStartTime: number;
    matchEndTime: number;
    venueUnitId: number;
    venueUnitNo: string;
    venueId: number;
    venueNo: string;
    physicalAreaId: number;
    matchResult: MatchResult;
    liveArena: boolean;
    liveClientType: string;
    statusTypeId: number;
    statusType: string;
    lastChangeDate: string;
    spectators: number | null;
    actualMatchDate: string;
    actualMatchStartTime: number | null;
    actualMatchEndTime: number | null;
    sportId: number;
};